Ported the unit tests to the new system.
