+++ tests/test_scripts.py
@@ -58,8 +58,6 @@ from scripts import (
     FixInvisibleWorksScript,
     LibraryInputScript,
     MockStdin,
-    OneClickDeltaScript,
-    OneClickImportScript, 
     OPDSImportScript,
     PatronInputScript,
     RunCollectionMonitorScript,
@@ -939,127 +937,6 @@ class TestAddClassificationScript(DatabaseTest):
         [classification] = identifier.classifications
         subject = classification.subject
         eq_("some random tag", subject.identifier)
-
-
-class TestOneClickImportScript(DatabaseTest):
-
-    def get_data(self, filename):
-        base_path = os.path.split(__file__)[0]
-        self.resource_path = os.path.join(base_path, "files", "oneclick")
-
-        # returns contents of sample file as string and as dict
-        path = os.path.join(self.resource_path, filename)
-        data = open(path).read()
-        return data, json.loads(data)
-
-    def test_import(self):
-        base_path = os.path.split(__file__)[0]
-        collection = MockOneClickAPI.mock_collection(self._db)
-        importer = OneClickImportScript(
-            collection, api_class=MockOneClickAPI, base_path=base_path
-        )
-        datastr, datadict = self.get_data("response_catalog_all_sample.json")
-        importer.api.queue_response(status_code=200, content=datastr)
-        importer.run()
-
-        # verify that we created Works, Editions, LicensePools
-        works = self._db.query(Work).all()
-        work_titles = [work.title for work in works]
-        expected_titles = ["Tricks", "Emperor Mage: The Immortals", 
-            "In-Flight Russian", "Road, The", "Private Patient, The", 
-            "Year of Magical Thinking, The", "Junkyard Bot: Robots Rule, Book 1, The", 
-            "Challenger Deep"]
-        eq_(set(expected_titles), set(work_titles))
-
-        # make sure we created some Editions
-        edition = Edition.for_foreign_id(self._db, DataSource.ONECLICK, Identifier.ONECLICK_ID, "9780062231727", create_if_not_exists=False)
-        assert(edition is not None)
-        edition = Edition.for_foreign_id(self._db, DataSource.ONECLICK, Identifier.ONECLICK_ID, "9781615730186", create_if_not_exists=False)
-        assert(edition is not None)
-
-        # make sure we created some LicensePools
-        pool, made_new = LicensePool.for_foreign_id(
-            self._db, DataSource.ONECLICK, Identifier.ONECLICK_ID,
-            "9780062231727", collection=collection
-        )
-        eq_(False, made_new)
-        pool, made_new = LicensePool.for_foreign_id(
-            self._db, DataSource.ONECLICK, Identifier.ONECLICK_ID,
-            "9781615730186", collection=collection
-        )
-        eq_(False, made_new)
-
-        # make sure there are 8 LicensePools
-        pools = self._db.query(LicensePool).all()
-        eq_(8, len(pools))
-
-
-class TestOneClickDeltaScript(DatabaseTest):
-
-    def get_data(self, filename):
-        base_path = os.path.split(__file__)[0]
-        self.resource_path = os.path.join(base_path, "files", "oneclick")
-
-        # returns contents of sample file as string and as dict
-        path = os.path.join(self.resource_path, filename)
-        data = open(path).read()
-        return data, json.loads(data)
-
-
-    def test_delta(self):
-        # First, load a collection.
-        base_path = os.path.split(__file__)[0]
-        collection = MockOneClickAPI.mock_collection(self._db)
-        importer = OneClickImportScript(
-            collection, api_class=MockOneClickAPI, base_path=base_path
-        )
-        datastr, datadict = self.get_data("response_catalog_all_sample.json")
-        importer.api.queue_response(status_code=200, content=datastr)
-        importer.run()
-
-        # set license numbers on test pool
-        pool, made_new = LicensePool.for_foreign_id(
-            self._db, DataSource.ONECLICK, Identifier.ONECLICK_ID,
-            "9781615730186", collection=collection
-        )
-        eq_(False, made_new)
-        pool.licenses_owned = 10
-        pool.licenses_available = 9
-        pool.licenses_reserved = 2
-        pool.patrons_in_hold_queue = 1
-
-        # now update that library with a sample delta            
-        delta_runner = OneClickDeltaScript(
-            collection, api_class=MockOneClickAPI, base_path=base_path
-        )
-        datastr, datadict = self.get_data("response_catalog_delta.json")
-        delta_runner.api.queue_response(status_code=200, content=datastr)
-        delta_runner.run()
-
-        # "Tricks" did not get deleted, but did get its pools set to "nope".
-        # "Emperor Mage: The Immortals" got new metadata.
-        works = self._db.query(Work).all()
-        work_titles = [work.title for work in works]
-        expected_titles = ["Tricks", "Emperor Mage: The Immortals", 
-            "In-Flight Russian", "Road, The", "Private Patient, The", 
-            "Year of Magical Thinking, The", "Junkyard Bot: Robots Rule, Book 1, The", 
-            "Challenger Deep"]
-        eq_(set(expected_titles), set(work_titles))
-
-        eq_("Tricks", pool.presentation_edition.title)
-        eq_(0, pool.licenses_owned)
-        eq_(0, pool.licenses_available)
-        eq_(0, pool.licenses_reserved)
-        eq_(0, pool.patrons_in_hold_queue)
-        assert (datetime.datetime.utcnow() - pool.last_checked) < datetime.timedelta(seconds=20)
-
-        # make sure we updated fields
-        edition = Edition.for_foreign_id(self._db, DataSource.ONECLICK, Identifier.ONECLICK_ID, "9781934180723", create_if_not_exists=False)
-        eq_("Recorded Books, Inc.", edition.publisher)
-
-        # make sure there are still 8 LicensePools
-        pools = self._db.query(LicensePool).all()
-        eq_(8, len(pools))
 
 
 class TestShowLibrariesScript(DatabaseTest):
